# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student Career Development and Placement Readiness System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job postings
      responses:
        "200":
          description: Job postings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobPosting"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPostingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"
  /jobs/{id}:
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPostingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/apply:
    post:
      operationId: applyToJob
      tags: [jobs]
      summary: Apply to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"

  # ─── Interviews ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    JobPosting:
      type: object
      required: [id, title, companyId, companyName, status]
      properties:
        id:
          type: integer
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        description:
          type: ["string", "null"]
        eligibilityCriteria:
          type: ["string", "null"]
        applicationDeadline:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed, upcoming]
        ctc:
          type: ["number", "null"]
        openings:
          type: ["integer", "null"]
        applicationCount:
          type: integer
        createdAt:
          type: string
    JobPostingInput:
      type: object
      required: [title, companyId]
      properties:
        title:
          type: string
        companyId:
          type: integer
        description:
          type: string
        eligibilityCriteria:
          type: string
        applicationDeadline:
          type: string
        status:
          type: string
          enum: [open, closed, upcoming]
        ctc:
          type: number
        openings:
          type: integer
    JobPostingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eligibilityCriteria:
          type: string
        applicationDeadline:
          type: string
        status:
          type: string
          enum: [open, closed, upcoming]
        ctc:
          type: number
        openings:
          type: integer
    JobApplication:
      type: object
      required: [id, jobId, studentId, status, appliedAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        studentId:
          type: integer
        status:
          type: string
          enum: [applied, shortlisted, interviewed, offered, rejected]
        appliedAt:
          type: string

    # ─── Interview ────────────────────────────────────────────────────────────
...