# lib/api-zod/src/generated/types/jobPosting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Student Career Development and Placement Readiness System API
 * OpenAPI spec version: 0.1.0
 */
import type { JobPostingStatus } from './jobPostingStatus';

export interface JobPosting {
  id: number;
  title: string;
  companyId: number;
  companyName: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  eligibilityCriteria?: string | null;
  /** @nullable */
  applicationDeadline?: string | null;
  status: JobPostingStatus;
  /** @nullable */
  ctc?: number | null;
  /** @nullable */
  openings?: number | null;
  applicationCount?: number;
  createdAt?: string;
}